SCP-1921: Metadata: tidying up
- add Mock module
- add MetadataUrl newtype
- explicit exports
